Main: 2nd try: fixes for a few -Wunused -Wconversion
svn:sync-xref-src-repo-rev: r108925

+++ src/VBox/Main/include/HGCMObjects.h
@@ -1,4 +1,4 @@
-/* $Id: HGCMObjects.h 62372 2016-07-20 17:25:41Z noreply@oracle.com $ */
+/* $Id: HGCMObjects.h 62379 2016-07-20 20:11:50Z noreply@oracle.com $ */
 /** @file
  *
  * HGCMObjects - Host-Guest Communication Manager objects header.
@@ -89,7 +89,7 @@ class HGCMObject
 
         uint32_t Handle()
         {
-            return m_core.AvlCore.Key;
+            return (uint32_t)m_core.AvlCore.Key;
         };
 
         HGCMOBJ_TYPE Type()
